Fix dot code generation for a single graph
Right now, everything works fine, except that the clusters are not
visualized in tree form, but rather unorderly all over the place.
If I leave out the graphs inside the clusters and only leave one
invisible node per cluster, then everything is perfectly aligned
in a tree.

It means that the subgraphs inside the clusters are not viewed as some
labels, but they are viewed as actual parts of the graph which is not
appropriate for my situation.